app/backend/edge-services/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
aiokafka
asyncpg
qdrant-client
numpy
Pillow
httpx
cryptography
prometheus-client
pynvml
boto3
orjson
onnxruntime
pytest
pytest-asyncio
